playground-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
transformers==4.36.2
torch==2.6.0
torchvision==0.21.0
pillow
librosa==0.10.1
numpy==1.26.4
python-multipart==0.0.6
dehug==0.2.2